dom/tests/mochitest/fetch/test_fetch_app_protocol.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1161288 - Support app:// origins on Fetch API</title>
  <script type='text/javascript' src='/tests/SimpleTest/SimpleTest.js'></script>
  <link rel='stylesheet' type='text/css' href='/tests/SimpleTest/test.css' />
</head>
<body onload='runTests()'>
<p id='display'></p>
<div id='content' style='display: none'></div>
<pre id='test'></pre>
<script class='testbody' type='application/javascript;version=1.7'>

SimpleTest.waitForExplicitFinish();

const appManifestURL =
  'http://mochi.test:8888/tests/dom/tests/mochitest/fetch/app-protocol/update.webapp';
let gApp;

function setup() {
  info('Setting up');
  return new Promise((resolve, reject) => {
    SpecialPowers.pushPrefEnv({'set': [
      ['dom.mozBrowserFramesEnabled', true]
    ]}, () => {
      SpecialPowers.pushPermissions([
        { 'type': 'webapps-manage', 'allow': 1, 'context': document },
        { 'type': 'browser', 'allow': 1, 'context': document },
        { 'type': 'embed-apps', 'allow': 1, 'context': document }
      ], () => {
        SpecialPowers.autoConfirmAppInstall(() => {
          SpecialPowers.autoConfirmAppUninstall(resolve);
        });
      });
    });
  });
}

function installApp() {
  return new Promise((resolve, reject) => {
    let req = navigator.mozApps.installPackage(appManifestURL);
    req.onsuccess = function() {
      gApp = req.result;
      is(req.result.manifestURL, appManifestURL, 'app installed');
      if (req.result.installState == 'installed') {
        is(req.result.installState, 'installed', 'app downloaded');
        resolve()
      } else {
        req.result.ondownloadapplied = function() {
          is(req.result.installState, 'installed', 'app downloaded');
          resolve();
        }
      }
    }
    req.onerror = reject;
  });
}

function launchApp() {
  return new Promise((resolve, reject) => {
    if (!gApp) {
      ok(false, 'No test application to launch');
      reject();
    }

    let expectedOkCount = 5;
    let okCount = 0;
    let iframe = document.createElement('iframe');
    iframe.setAttribute('mozbrowser', 'true');
    iframe.setAttribute('mozapp', gApp.manifestURL);
    let domParent = document.getElementById('container');
    iframe.addEventListener('mozbrowsershowmodalprompt', function listener(e) {
      let message = e.detail.message;
      info(message);
      if (/OK/.exec(message)) {
        ok(true, "Message from app: " + message);
        okCount++;
      } else if (/KO/.exec(message)) {
        ok(false, "Message from app: " + message);
      } else if (/DONE/.exec(message)) {
        ok(true, "Messaging from app complete");
        iframe.removeEventListener('mozbrowsershowmodalprompt', listener);
        domParent.removeChild(iframe);
        ok(okCount == expectedOkCount, "Everything's fine");
        resolve();
      }
    }, false);
    domParent.appendChild(iframe);
    SpecialPowers.wrap(iframe.contentWindow).location =
      gApp.origin + gApp.manifest.launch_path;
  });
}

function uninstallApp() {
  return new Promise((resolve, reject) => {
    if (!gApp) {
      return reject();
    }
    let req = navigator.mozApps.mgmt.uninstall(gApp);
    req.onsuccess = resolve;
    req.onerror = reject;
  });
}

function runTests() {
  setup()
    .then(installApp)
    .then(launchApp)
    .then(uninstallApp)
    .then(SimpleTest.finish)
    .catch((e) => {
      ok(false, 'Unexpected error ' + e);
      SimpleTest.finish();
    });
}

</script>
<div id='container'></div>
</body>
</html>
